Git: Fix ANSI escape sequence bug in log output
A move to libsmartcols in the column tool introduced a bug that disables
the output of colors via ANSI escape sequences.
Removing the -t --table option fixes this, but may introduce issues with
formatting, if an empty field should appear.

Update the githelper script to a newer version from Bernhardt's repo.

Remove color settings from .gitconfig-common and reindented the file